+++ daemon/manager/manager.go
@@ -28,7 +28,6 @@ import (
 	"github.com/spf13/pflag"
 
 	"kmesh.net/kmesh/daemon/manager/dump"
-	logcmd "kmesh.net/kmesh/daemon/manager/log"
 	"kmesh.net/kmesh/daemon/manager/uninstall"
 	"kmesh.net/kmesh/daemon/manager/version"
 	"kmesh.net/kmesh/daemon/options"
@@ -68,7 +67,6 @@ func NewCommand() *cobra.Command {
 	// add sub commands
 	cmd.AddCommand(version.NewCmd())
 	cmd.AddCommand(dump.NewCmd())
-	cmd.AddCommand(logcmd.NewCmd())
 	cmd.AddCommand(uninstall.NewCmd())
 
 	return cmd
